Successfully wrote out bias and ghost matrix information as ENVI spectral library.
A routine in order to read the spectral library and separate out
the bias and matrix so that it can be applied through Emmett's
matrix inversion routine.

The Flexible spectral library data structure may need to be incorporated
in order to allow sorting of the bias values to allow the selection
and concatenation to be made.

; Please enter the commit message for your changes. Lines starting
; with '#' will be ignored, and an empty message aborts the commit.
;
; Committer: U-homestar\rvrpci <rvrpci@homestar.(none)>
;
; On branch master
; Changes to be committed:
;   (use "git reset HEAD <file>..." to unstage)
;
;	modified:   GOES-B11-L8-MODIS-geo-bias-water-flicker.hdr
;	modified:   GOES-B11-L8-MODIS-geo-bias-water-sampled.hdr
;	modified:   GOES-B11-L8-MODIS-geo-bias-water-unsampled.hdr
;	modified:   GOES_ghost_matrix.hdr
;	modified:   L8_ghost_matrix.hdr
;	modified:   fly_tirs_for_ghost_matrix.pro
;	modified:   write_bias_ghost_matrix_as_spectral_library.pro
;